Google export
PiperOrigin-RevId: 309836740
