+++ epi_forecast_stat_mech/tests/test_high_level.py
@@ -4,7 +4,6 @@
 from absl.testing import absltest
 
 from epi_forecast_stat_mech import high_level
-from epi_forecast_stat_mech import rtlive
 from epi_forecast_stat_mech import sir_sim
 
 from jax.config import config
@@ -57,7 +56,21 @@ class TestHighLevel(absltest.TestCase):
     num_samples = 11
 
     data = create_synthetic_dataset(num_epidemics=50, num_time_steps=100)
-    estimator = rtlive.RtLiveEstimator(gamma=1.0).fit(data)
+    estimator = high_level.RtLiveEstimator(gamma=1.0).fit(data)
+
+    predictions = estimator.predict(prediction_length, num_samples)
+    self.assertCountEqual(['location', 'sample', 'time'], predictions.dims)
+    self.assertLen(predictions.time, prediction_length)
+    np.testing.assert_array_equal(data.location, predictions.location)
+    self.assertLen(predictions.sample, num_samples)
+
+  def test_SparseEstimator(self):
+    """Verify we can fit and predict from SparseEstimator."""
+    prediction_length = 10
+    num_samples = 11
+
+    data = create_synthetic_dataset(num_epidemics=50, num_time_steps=100)
+    estimator = high_level.StatMechEstimator().fit(data, train_steps=1000)
 
     predictions = estimator.predict(prediction_length, num_samples)
     self.assertCountEqual(['location', 'sample', 'time'], predictions.dims)
